Moved AuthRequest to DTOs
